feat: Add script for running genome guided Trinity assembly
This commit adds a new script, run_trinity.sh, which is used to run a genome guided Trinity assembly on a merged bam file. The script takes several arguments, including the path to the merged bam file, the maximum intron size, memory allocation, and CPU usage. The output of the assembly is saved in the "trinity" directory.